requirements.txt
pandas==1.0.1
streamlit==0.78.0
numpy==1.18.1
google==3.0.0
Pillow==8.1.2
